This commit removes the BibTeX parser in bibtex.py, and replaces it with a dependency on the bibtexparser module from PyPI.
The rationale for this change is that the original parser got stuck
in a regex on my 1.1MB test bib-file, and seemed generally to scale
quite badly with file size. In addition, the bibtexparser module seems
to solve the problem of parsing BibTeX nicely, therefore I think it's
best to reuse that solution.

With the new parser, I can import my 1.1MB bib-file, containing
about 1460 references, in about 16 seconds.